Move graphql_value! macro into separate module
